Refactor: adapt to btckApi changes and implement rust idioms
Breaking changes:
- Replace BlockUndo with hierarchical
  BlockSpentOutputs/TransactionSpentOutputs/Coin design
- Remove get_ prefixes from all methods following rust naming
  conventions
- Introduce RefType<T, Owner> for zero-copy borrowing

Improvements:
- Add RefType wrapper for memory-safe borrowed references with
  compile-time lifetime guarantees
- Update internal bindings from kernel_ to btck_ prefixed C API
- Add input validation to verify function
- Add transaction and block iteration methods (output_count,
  transaction_count, etc.)
- Implement Clone traits using C copy functions
- Leverage C Library's smart ownershop tracking (m_owned flags) for safe
  borrowing
- Updated and improved documentation

+++ fuzz/fuzz_targets/fuzz_target_verify.rs
@@ -16,7 +16,7 @@ pub struct VerifyInput {
     pub script_pubkey: Vec<u8>,
     pub amount: Option<i64>,
     pub tx_to: Vec<u8>,
-    pub input_index: u32,
+    pub input_index: usize,
     pub flags: Option<u32>,
     pub spent_outputs: Vec<UtxoWrapper>,
 }
